Create SnakeGame.cpp
This project aims to bring the fun and simplicity of snake game with some new features. It will
include computer controlled intelligent opponents whose aim will be to challenge the human
players. It will also have the multiplayer feature that will allow more than one players to play
the game over a network.
This project explores a new dimension in the traditional snake game to make it more
interesting and challenging. The simplicity of this game makes it an ideal candidate for a
minor project as we can focus on advanced topics like multiplayer functionality and
implementation of computer controlled intelligent opponents.