Create CQuizeGames.cpp
The sole intention behind the consideration of this Project is to generate and manage a simple database for question. This project is developed considering “QUIZ” informationkeeping context of the customer in mind. Here, data is stored in a binary file by “QUESTION” and thisdatabase is basically used  as MASTER file